+++ qtdeclarative/src/qml/qml/qqmlvme_p.h
@@ -1,5 +1,6 @@
 // Copyright (C) 2016 The Qt Company Ltd.
 // SPDX-License-Identifier: LicenseRef-Qt-Commercial OR LGPL-3.0-only OR GPL-2.0-only OR GPL-3.0-only
+// Qt-Security score:significant
 
 #ifndef QQMLVME_P_H
 #define QQMLVME_P_H
@@ -25,7 +26,6 @@
 #include <QtCore/qtypeinfo.h>
 
 #include <private/qqmlengine_p.h>
-#include <private/qfinitestack_p.h>
 
 #include <atomic>
 
@@ -78,7 +78,7 @@ public:
     bool isOK() const;
 
 private:
-    int m_objectCount;
+    size_t m_objectCount;
     QQmlGuard<QObject> *m_objects;
     int m_contextCount;
     QQmlGuardedContextData *m_contexts;
